+++ navigation-extender-runtime/src/main/java/com/foursoft/xml/ExtendedUnmarshaller.java
@@ -28,13 +28,12 @@ package com.foursoft.xml;
 import com.foursoft.xml.annotations.XmlBackReference;
 import com.foursoft.xml.annotations.XmlParent;
 import com.foursoft.xml.postprocessing.*;
-import com.sun.xml.bind.v2.util.XmlFactory;
+import jakarta.xml.bind.*;
 import org.xml.sax.InputSource;
 import org.xml.sax.SAXException;
 import org.xml.sax.XMLReader;
 
 import javax.xml.XMLConstants;
-import javax.xml.bind.*;
 import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.parsers.SAXParser;
 import javax.xml.parsers.SAXParserFactory;
@@ -46,7 +45,6 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.function.Consumer;
 import java.util.function.Function;
-import java.util.logging.Level;
 
 /**
  * Provides extended unmarshalling capabilities for a JAXB model. Capabilities
@@ -83,7 +81,7 @@ public class ExtendedUnmarshaller<R, I> {
     public ExtendedUnmarshaller(final Class<R> rootElement) throws JAXBException {
         this.rootElement = rootElement;
         final String packageName = rootElement.getPackage().getName();
-        final JAXBContext context = JaxbContextFactory.initializeContext(packageName, this.getClass().getClassLoader());
+        final JAXBContext context = JaxbContextFactory.initializeContext(packageName, getClass().getClassLoader());
         postProcessorRegistry = new ModelPostProcessorRegistry(packageName);
         unmarshaller = context.createUnmarshaller();
     }
@@ -165,7 +163,7 @@ public class ExtendedUnmarshaller<R, I> {
 
         unmarshaller.setListener(modelPostProcessorManager);
 
-        final R root = rootElement.cast(unmarshaller.unmarshal(new SAXSource(getXMLReader(),new InputSource(resource))));
+        final R root = rootElement.cast(unmarshaller.unmarshal(new SAXSource(getXMLReader(), new InputSource(resource))));
 
         modelPostProcessorManager.doPostProcessing();
 
@@ -193,6 +191,33 @@ public class ExtendedUnmarshaller<R, I> {
         postProcessorRegistry.clearStateOfPostProcessors();
     }
 
+    /**
+     * Obtains a configured XMLReader.
+     * <p>
+     * {@link Unmarshaller} is not re-entrant, so we will
+     * only use one instance of XMLReader.
+     * <p>
+     * Overriden in order to fix potential security issue.
+     * See jakarta.xml.bind.helpers.AbstractUnmarshallerImpl#getXMLReader()
+     */
+    protected XMLReader getXMLReader() throws JAXBException {
+        if (reader == null) {
+            try {
+                final SAXParserFactory factory = SAXParserFactory.newInstance();
+                factory.setNamespaceAware(true);
+                factory.setFeature(XMLConstants.FEATURE_SECURE_PROCESSING, true);
+                factory.setValidating(false);
+                final SAXParser parser = factory.newSAXParser();
+                parser.setProperty(XMLConstants.ACCESS_EXTERNAL_DTD, "");
+                parser.setProperty(XMLConstants.ACCESS_EXTERNAL_SCHEMA, "");
+                reader = parser.getXMLReader();
+            } catch (final ParserConfigurationException | SAXException e) {
+                throw new JAXBException(e);
+            }
+        }
+        return reader;
+    }
+
     public static class InterceptEventHandler implements ValidationEventHandler {
         private final ValidationEventHandler originalEventHandler;
         private Consumer<ValidationEvent> eventConsumer;
@@ -205,7 +230,8 @@ public class ExtendedUnmarshaller<R, I> {
             this.eventConsumer = eventConsumer;
         }
 
-        @Override public boolean handleEvent(final ValidationEvent event) {
+        @Override
+        public boolean handleEvent(final ValidationEvent event) {
             eventConsumer.accept(event);
             return originalEventHandler.handleEvent(event);
         }
@@ -214,34 +240,6 @@ public class ExtendedUnmarshaller<R, I> {
             Objects.requireNonNull(eventConsumer);
             this.eventConsumer = eventConsumer;
         }
-    }
-
-    /**
-     * Obtains a configured XMLReader.
-     *
-     * {@link Unmarshaller} is not re-entrant, so we will
-     * only use one instance of XMLReader.
-     *
-     * Overriden in order to fix potential security issue.
-     * See javax.xml.bind.helpers.AbstractUnmarshallerImpl#getXMLReader()
-     *
-     */
-    protected XMLReader getXMLReader() throws JAXBException {
-        if (reader == null) {
-            try {
-                SAXParserFactory factory = SAXParserFactory.newInstance();
-                factory.setNamespaceAware(true);
-                factory.setFeature(XMLConstants.FEATURE_SECURE_PROCESSING, true);
-                factory.setValidating(false);
-                SAXParser parser = factory.newSAXParser();
-                parser.setProperty(XMLConstants.ACCESS_EXTERNAL_DTD,"");
-                parser.setProperty(XMLConstants.ACCESS_EXTERNAL_SCHEMA,"");
-                reader = parser.getXMLReader();
-            } catch (ParserConfigurationException|SAXException e) {
-                throw new JAXBException(e);
-            }
-        }
-        return reader;
     }
 
 }
