+++ navigation-extender-runtime/src/main/java/com/foursoft/xml/JaxbContextFactory.java
@@ -25,8 +25,9 @@
  */
 package com.foursoft.xml;
 
-import javax.xml.bind.JAXBContext;
-import javax.xml.bind.JAXBException;
+import jakarta.xml.bind.JAXBContext;
+import jakarta.xml.bind.JAXBException;
+
 import java.util.HashMap;
 import java.util.Map;
