+++ navigation-extender-runtime/src/main/java/com/foursoft/xml/io/read/XMLReader.java
@@ -29,11 +29,11 @@ import com.foursoft.xml.ExtendedUnmarshaller;
 import com.foursoft.xml.JaxbModel;
 import com.foursoft.xml.io.utils.ValidationEventLogger;
 import com.foursoft.xml.io.utils.XMLIOException;
+import jakarta.xml.bind.JAXBException;
+import jakarta.xml.bind.ValidationEvent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.xml.bind.JAXBException;
-import javax.xml.bind.ValidationEvent;
 import java.io.BufferedInputStream;
 import java.io.IOException;
 import java.io.InputStream;
