+++ navigation-extender-runtime/src/main/java/com/foursoft/xml/postprocessing/ModelPostProcessorManager.java
@@ -25,7 +25,8 @@
  */
 package com.foursoft.xml.postprocessing;
 
-import javax.xml.bind.Unmarshaller;
+import jakarta.xml.bind.Unmarshaller;
+
 import java.util.LinkedList;
 import java.util.List;
