+++ navigation-extender-runtime/src/test/java/com/foursoft/xml/io/validation/LogErrorsTest.java
@@ -43,10 +43,9 @@ class LogErrorsTest {
         final XMLValidation xmlValidation = getXmlValidation();
 
         final String content = new String(
-                Files.readAllBytes(TestData.VALIDATE_BASE_PATH.resolve(TestData.ERROR_TEST_XML)));
-
+                Files.readAllBytes(TestData.VALIDATE_BASE_PATH_SRC.resolve(TestData.ERROR_TEST_XML)));
         final Collection<ErrorLocation> errors = xmlValidation.validateXML(content,
-                                                                           StandardCharsets.UTF_8);
+                StandardCharsets.UTF_8);
         final String errorString = LogErrors.annotateXMLContent(content, errors);
         // line 21 contains a duplicate key
         assertThat(errorString).contains("21: ERROR     <ChildB id=\"id_8\">");
