+++ navigation-extender-runtime/src/test/java/com/foursoft/xml/model/AbstractBase.java
@@ -1,8 +1,8 @@
 /*-
  * ========================LICENSE_START=================================
- * xml-runtime
+ * navigation-extender-runtime
  * %%
- * Copyright (C) 2019 4Soft GmbH
+ * Copyright (C) 2019 - 2020 4Soft GmbH
  * %%
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -23,37 +23,42 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.xml.postprocessing;
+package com.foursoft.xml.model;
 
-import jakarta.xml.bind.Unmarshaller;
+import jakarta.xml.bind.annotation.*;
+import jakarta.xml.bind.annotation.adapters.CollapsedStringAdapter;
+import jakarta.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
 
-public interface ModelPostProcessor {
+import java.io.Serializable;
 
-    Class<?> getClassToHandle();
+@XmlAccessorType(XmlAccessType.FIELD)
+@XmlType(name = "AbstractBase", propOrder = {})
+@XmlSeeAlso({Root.class, ChildA.class, ChildB.class})
+public abstract class AbstractBase implements Serializable {
+
+    private final static long serialVersionUID = 1L;
+    @XmlAttribute(name = "id", required = true)
+    @XmlJavaTypeAdapter(CollapsedStringAdapter.class)
+    @XmlID
+    @XmlSchemaType(name = "ID")
+    protected String xmlId;
 
     /**
-     * Called during the unmarshalling the JAXB unmarshalling process, after the unmarshalling of <tt>target</tt> has
-     * been completed. IDREF relations are not yet initialized at this point of the unmarshalling process.
+     * Ruft den Wert der xmlId-Eigenschaft ab.
      *
-     * @param target non-null instance of JAXB mapped class prior to unmarshalling into it.
-     * @param parent instance of JAXB mapped class that will reference
-     *               <tt>target</tt>. <tt>null</tt> when <tt>target</tt> is root
-     *               element.
-     * @see Unmarshaller.Listener#afterUnmarshal(Object, Object)
+     * @return possible object is {@link String }
      */
-    void afterUnmarshalling(Object target, Object parent);
+    public String getXmlId() {
+        return xmlId;
+    }
 
-    /**
-     * Called after the unmarshalling process of the whole model has been completed.
-     *
-     * @param target the object co call
-     */
-    void afterUnmarshallingCompleted(Object target);
 
-    /**
-     * Clear the state of the {@link ModelPostProcessor} to allow garbage collection and reuse during multiple
-     * unmarshallings.
-     */
-    void clearState();
+    public void setXmlId(final String value) {
+        xmlId = value;
+    }
 
+    @Override
+    public String toString() {
+        return getClass().getSimpleName() + "[" + getXmlId() + "]";
+    }
 }
