+++ navigation-extender-runtime/src/test/java/com/foursoft/xml/model/ChildA.java
@@ -10,10 +10,10 @@
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -23,53 +23,52 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.test.model;
+package com.foursoft.xml.model;
+
+import com.foursoft.xml.annotations.XmlBackReference;
+import com.foursoft.xml.annotations.XmlParent;
+import jakarta.xml.bind.annotation.*;
 
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
-import javax.xml.bind.annotation.XmlType;
-
 @XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "Root", propOrder = { "attribute", "childA", "childB" })
-@XmlRootElement(name = "Root")
-public class Root extends AbstractBase implements Serializable {
+@XmlType(name = "ChildA", propOrder = {"anotherAttribute", "referencedChildB"})
+public class ChildA extends AbstractBase implements Serializable {
 
-	private final static long serialVersionUID = 1L;
+    private final static long serialVersionUID = 1L;
 
-	@XmlElement(name = "attribute", required = true)
-	protected String attribute;
+    @XmlElement(name = "anotherAttribute", required = true)
+    protected String anotherAttribute;
 
-	@XmlElement(name = "ChildA")
-	protected List<ChildA> childA;
-	@XmlElement(name = "ChildB")
-	protected List<ChildB> childB;
+    @XmlList
+    @XmlElement(name = "ReferencedChildB", type = java.lang.Object.class)
+    @XmlIDREF
+    @XmlSchemaType(name = "IDREFS")
+    @XmlBackReference(destinationField = "reverseChildA")
+    protected List<ChildB> referencedChildB;
+    @XmlTransient
+    @XmlParent
+    private Root root;
 
-	public String getAttributeA() {
-		return attribute;
-	}
+    public String getAnotherAttribute() {
+        return anotherAttribute;
+    }
 
-	public void setAttribute(final String attribute) {
-		this.attribute = attribute;
-	}
+    public void setAnotherAttribute(final String anotherAttribute) {
+        this.anotherAttribute = anotherAttribute;
+    }
 
-	public List<ChildA> getChildA() {
-		if (childA == null) {
-			childA = new ArrayList<>();
-		}
-		return this.childA;
-	}
+    public List<ChildB> getReferencedChildB() {
+        if (referencedChildB == null) {
+            referencedChildB = new ArrayList<>();
+        }
+        return referencedChildB;
+    }
 
-	public List<ChildB> getChildB() {
-		if (childB == null) {
-			childB = new ArrayList<>();
-		}
-		return this.childB;
-	}
+    public Root getRoot() {
+        return root;
+    }
 
 }
