+++ navigation-extender-runtime/src/test/java/com/foursoft/xml/model/ChildB.java
@@ -10,10 +10,10 @@
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -23,49 +23,41 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.test.model;
+package com.foursoft.xml.model;
+
+import com.foursoft.xml.annotations.XmlParent;
+import jakarta.xml.bind.annotation.*;
 
 import java.io.Serializable;
-
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlAttribute;
-import javax.xml.bind.annotation.XmlID;
-import javax.xml.bind.annotation.XmlSchemaType;
-import javax.xml.bind.annotation.XmlSeeAlso;
-import javax.xml.bind.annotation.XmlType;
-import javax.xml.bind.annotation.adapters.CollapsedStringAdapter;
-import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
+import java.util.HashSet;
+import java.util.Set;
 
 @XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "AbstractBase", propOrder = {})
-@XmlSeeAlso({ Root.class, ChildA.class, ChildB.class })
-public abstract class AbstractBase implements Serializable {
+@XmlType(name = "ChildB", propOrder = {"xyz",})
+public class ChildB extends AbstractBase implements Serializable {
 
-	private final static long serialVersionUID = 1L;
-	@XmlAttribute(name = "id", required = true)
-	@XmlJavaTypeAdapter(CollapsedStringAdapter.class)
-	@XmlID
-	@XmlSchemaType(name = "ID")
-	protected String xmlId;
+    private final static long serialVersionUID = 1L;
+    @XmlTransient
+    private final Set<ChildA> reverseChildA = new HashSet<>();
+    @XmlElement(name = "xyz", required = true)
+    protected String xyz;
+    @XmlTransient
+    @XmlParent
+    private Root root;
 
-	/**
-	 * Ruft den Wert der xmlId-Eigenschaft ab.
-	 * 
-	 * @return possible object is {@link String }
-	 * 
-	 */
-	public String getXmlId() {
-		return xmlId;
-	}
+    public String getXyz() {
+        return xyz;
+    }
 
+    public void setXyz(final String xyz) {
+        this.xyz = xyz;
+    }
 
-	public void setXmlId(final String value) {
-		this.xmlId = value;
-	}
+    public Root getRoot() {
+        return root;
+    }
 
-	@Override
-	public String toString() {
-		return this.getClass().getSimpleName() + "[" + this.getXmlId() + "]";
-	}
+    public Set<ChildA> getReverseChildA() {
+        return reverseChildA;
+    }
 }
