+++ navigation-extender-runtime/src/test/java/com/foursoft/xml/model/ObjectFactory.java
@@ -23,9 +23,9 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.test.model;
+package com.foursoft.xml.model;
 
-import javax.xml.bind.annotation.XmlRegistry;
+import jakarta.xml.bind.annotation.XmlRegistry;
 
 /**
  * An ObjectFactory allows you to programatically construct new instances of the
