+++ navigation-extender-runtime/src/test/java/com/foursoft/xml/model/Root.java
@@ -10,10 +10,10 @@
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -23,47 +23,49 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.test.model;
+package com.foursoft.xml.model;
+
+import jakarta.xml.bind.annotation.*;
 
 import java.io.Serializable;
-import java.util.HashSet;
-import java.util.Set;
-
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlTransient;
-import javax.xml.bind.annotation.XmlType;
-
-import com.foursoft.xml.annotations.XmlParent;
+import java.util.ArrayList;
+import java.util.List;
 
 @XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "ChildB", propOrder = { "xyz", })
-public class ChildB extends AbstractBase implements Serializable {
+@XmlType(name = "Root", propOrder = {"attribute", "childA", "childB"})
+@XmlRootElement(name = "Root")
+public class Root extends AbstractBase implements Serializable {
 
-	private final static long serialVersionUID = 1L;
+    private final static long serialVersionUID = 1L;
 
-	@XmlElement(name = "xyz", required = true)
-	protected String xyz;
-	@XmlTransient
-	private final Set<ChildA> reverseChildA = new HashSet<>();
-	@XmlTransient
-	@XmlParent
-	private Root root;
+    @XmlElement(name = "attribute", required = true)
+    protected String attribute;
 
-	public String getXyz() {
-		return xyz;
-	}
+    @XmlElement(name = "ChildA")
+    protected List<ChildA> childA;
+    @XmlElement(name = "ChildB")
+    protected List<ChildB> childB;
 
-	public void setXyz(final String xyz) {
-		this.xyz = xyz;
-	}
+    public String getAttributeA() {
+        return attribute;
+    }
 
-	public Root getRoot() {
-		return root;
-	}
+    public void setAttribute(final String attribute) {
+        this.attribute = attribute;
+    }
 
-	public Set<ChildA> getReverseChildA() {
-		return reverseChildA;
-	}
+    public List<ChildA> getChildA() {
+        if (childA == null) {
+            childA = new ArrayList<>();
+        }
+        return childA;
+    }
+
+    public List<ChildB> getChildB() {
+        if (childB == null) {
+            childB = new ArrayList<>();
+        }
+        return childB;
+    }
+
 }
