+++ navigation-extender/src/main/java/com/foursoft/xml/xjc/plugin/EmptyListHandlerPlugin.java
@@ -25,7 +25,7 @@
  */
 package com.foursoft.xml.xjc.plugin;
 
-import javax.xml.bind.Marshaller;
+import jakarta.xml.bind.Marshaller;
 
 import org.xml.sax.ErrorHandler;
 import org.xml.sax.SAXException;
