+++ navigation-extender/src/main/java/com/foursoft/xml/xjc/plugin/PropertyBuilder.java
@@ -25,7 +25,7 @@
  */
 package com.foursoft.xml.xjc.plugin;
 
-import javax.xml.bind.annotation.XmlTransient;
+import jakarta.xml.bind.annotation.XmlTransient;
 
 import com.sun.codemodel.JCodeModel;
 import com.sun.codemodel.JDefinedClass;
