+++ navigation-extender/src/main/java/module-info.java
@@ -1,8 +1,8 @@
 /*-
  * ========================LICENSE_START=================================
- * xml-runtime
+ * navigation-extender
  * %%
- * Copyright (C) 2019 4Soft GmbH
+ * Copyright (C) 2019 - 2022 4Soft GmbH
  * %%
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -23,37 +23,7 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.xml.postprocessing;
-
-import jakarta.xml.bind.Unmarshaller;
-
-public interface ModelPostProcessor {
-
-    Class<?> getClassToHandle();
-
-    /**
-     * Called during the unmarshalling the JAXB unmarshalling process, after the unmarshalling of <tt>target</tt> has
-     * been completed. IDREF relations are not yet initialized at this point of the unmarshalling process.
-     *
-     * @param target non-null instance of JAXB mapped class prior to unmarshalling into it.
-     * @param parent instance of JAXB mapped class that will reference
-     *               <tt>target</tt>. <tt>null</tt> when <tt>target</tt> is root
-     *               element.
-     * @see Unmarshaller.Listener#afterUnmarshal(Object, Object)
-     */
-    void afterUnmarshalling(Object target, Object parent);
-
-    /**
-     * Called after the unmarshalling process of the whole model has been completed.
-     *
-     * @param target the object co call
-     */
-    void afterUnmarshallingCompleted(Object target);
-
-    /**
-     * Clear the state of the {@link ModelPostProcessor} to allow garbage collection and reuse during multiple
-     * unmarshallings.
-     */
-    void clearState();
-
+module navigation.extender {
+    requires com.sun.tools.xjc;
+    requires transitive navigation.extender.runtime;
 }
