+++ pom.xml
@@ -6,7 +6,7 @@
     <parent>
         <groupId>com.foursoft.jaxb2</groupId>
         <artifactId>jaxb-parent</artifactId>
-        <version>0.0.9</version>
+        <version>0.0.10-SNAPSHOT</version>
     </parent>
 
     <groupId>com.foursoft.jaxb2</groupId>
@@ -21,7 +21,7 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
 
         <!-- Dependencies -->
